Extract 'add another course' actions from CourseChoicesController
This is a distinct phase of the course choice flow, prompting the user
to repeat the process after having successfully added a course. Place
these actions in their own controller, namespaced under the
CourseChoices module.
